Initialise constants for the BeaconChain class
Constants are defined in the specification. I've given them default visibility:
since they are `static final` there is no risk to encapsulation, so they don't
need to be private.

The spec defines constants in powers of 2 (mostly), so I've kept that convention.

GENESIS_TIME is set when the BeacinChain object is instantiatied.