Implement updateProposerRandaoLayer
* Rename updateProposerRandaoSkips to updateProposerRandaoLayer
* Add validator_registry and shard_committees_at_slots initialization to
the default constructor of BeaconState to enable slot processing action
to be taken when Artemis is ran, without giving an error due to null variables.
* Create a public getBeacon_proposer_index method to BeaconState to be
able to call the method from the StateTransition module.

NOTES:
- I haven't been able to test it yet, due to failure to think of what values
to initialize the validator_registry and shard_committees_at_slots with.
- If UpdateProposerRandaoLayer can become a method of BeaconState
instead of StateTransition, there wouldn't be a public method
implemented for getBeacon_proposer_index. It's private counterpart
get_beacon_proposer_index could be used.