[WIP] add libraries required to get swagger displaying.
Restructured GenesisTimeHandler to allow custom Summary, Description, tags (grouping)

for some reason the licence string for MIT was not the same as other libraries, updated the licence check script.

Refactored BeaconRestApi class to be singleton, which allows us to follow the standard context pattern when we build the API endpoints.