make changes ready to merge on master
[skip ci] as we don't want the branch push to update the real doc
the update will be triggered when merged on master.
The tag part (creating a version when a tag is created) is not yet
available. Only the latest will be built automatically for the
moment until this tag CI part is solved. Until then at least we will
have the master updating the latest doc version. If needed a named
version can be generated manually until we figure this out.