Implement a no-op database for local development.
The database directory is still created with the db.version but no chain data is ever stored.
Useful for testing in isolation of RocksDB but also for local runs during development to avoid having to delete the data dir before each run.
Have to opt-in to using it with `--Xdata-storage-create-db-version=noop`