add a command line interface to send voluntary exits for a validator.
The option is hidden,
`teku voluntary-exit --beacon-node-api-endpoint URL --validator-keys KEY:PASS --epoch EPOCH`

The validator key and password file are used to load the validator, so that the exit can be signed.

Care should be taken, if the entire validator keys folder and password folder is passed, exits will be generated for all keys.

If the validation from the remote host fails, the failure message is too generic to be useful currently, so the program only reports that the exit submission failed.

Note: the epoch must be the current or a previous epoch, or a validation error will be received, and the exit won't be processed.

fixes #2461

Signed-off-by: Paul Harris <paul.harris@consensys.net>