default the epoch for the voluntary-exit command to current epoch
Make the --epoch argument of voluntary-exit optional, and default to the current epoch, as future epochs aren't currently supported anyway.

Updated some of the validations to be debug level instead of trace, and added the validator index to the log line, but these would appear in the beacon node log, because that's when they run. Still, it would be helpful if you are looking for why a validator can't be exited, the 'will never pass' message isn't strictly accurate - for example if you're attempting to exit too early.

Added `--confirmation-enabled` to allow skipping user input to exit validators (--confirmation-enabled=false), but the greatest care should be exercised, as exits will be immediately generated with no prompt.

Signed-off-by: Paul Harris <paul.harris@consensys.net>