Avoid BigInteger multiplication when calculating base rewards.
Previously UInt64.times switched to using BigInteger to perform multiplication when either multiplicand is bigger than the square root of UInt64.MAX_VALUE to provide overflow protection.
Epoch transition multiplies the sum of effective balances by 64 and the effective balances were greater than the square root value, causing the more expensive path to be taken.
Add an additional check that if one multiplicand is 64 or less and the other is less than MAX_VALUE / 64, use the fast path as the result is known to not overflow.