ensure startup fails if keys cannot be loaded.
When a lock file is present and it's not safe to remove it, validators failed to load, but the client would sit in a state where no validators were loaded and not exit. The node ends up running but being non functional, which is undesirable.

This looks to be new behavior since keys were pushed to load asynchronously, resulting errors were not propagated up to the main controller.

fixes #3438

Signed-off-by: Paul Harris <paul.harris@consensys.net>