Apply votes to store independently of disk storage (#3600)
Votes are now applied in memory first and then the changes are sent to disk. This ensures that the in-memory store stays in sync even if the disk storage encounters a problem. At startup, the weightings in ProtoArray are rebuilt from the stored votes so even if some updates to votes fail to store to disk, when reloaded the ProtoArray is still guaranteed to be in sync with them.