Load the deposit contract from constants if it hasn't been set
If the network configuration is missing a deposit contract address, attempt to load it from the constants.

Because all constants files have deposit contract set, the test for not having contract address has been removed, but the error has been left in case someone specifies a custom constants file that has the contract missing.

fixes #3723

Signed-off-by: Paul Harris <paul.harris@consensys.net>