Update --ws-checkpoint to accept a URL
`--ws-checkpoint` now accepts either root:epoch, or a URL containing key `ws_checkpoint` (formatted as root:epoch) as part of a json payload.

If the checkpoint passed starts with '0x', then it will be loaded as a literal checkpoint reference, otherwise it will attempt to load from url.

If the checkpoint doesn't start in 0x, and it doesnt load from URL, teku will fail with a CLI error.

Fixes #3531

Signed-off-by: Paul Harris <paul.harris@consensys.net>