Calculate epoch based on spec from http endpoint for exits
The endpoint provides the full spec, but the spec builder takes a map of string to object so the output of the api wasn't fully compatible. Ideally I could just load from the http response rather than using stored configuration, but at least the name will provide the spec to load.

fixes #3862

Signed-off-by: Paul Harris <paul.harris@consensys.net>