Update remote validator client to handle altair block creation
Added a hidden cli option to direct the validator client to use `eth/v2/validator/blocks/:slot`, so that the version field in the response can be used to determine which type of block should be deserialized.

The `eth/v1/validator/blocks/:slot` endpoint does return an altair block but its less simple to decode, which ultimately stopped the remote validator from participating in block production.

The block production v2 endpoint is part of the standard spec, but we need to be careful switching to it, so for now this option is hidden and disabled by default.

For deserialization to work, a custom deserializer has been added for the GetNewBlockResponseV2 object.

it appears theres still functionality broken for remote validators, but this is a good start.

fixes #3974

Signed-off-by: Paul Harris <paul.harris@consensys.net>