Reduce complexity and object creation in VoteTracker
There are a lot of VoteTracker instances created and stored, to improve performance we already use a custom serializer when storing to the database so don't need them to use the SSZ framework.  Making them a POJO simplifies the code, but also avoids the need for boxing of the Bytes32 and UInt64 values they contain so reduces GC load as they are created and destroyed.