Fix missing default values for boolean options in help
CLI command classes such as `BeaconNodeCommand` are already configured to display the default values of their options in the usage help message.
However, this setting does not apply to boolean-typed options, and thus their default values are currently missing from the help message (issue #1746).
This commit addresses this issue by explicitly enabling the visibility of the default values of each boolean option that is visible to the user.