Test visibility of defaults for boolean CLI options
As discussed in #1746 all CLI boolean options should display their default values in the usage help message. This requires that the visibility of each boolean option is specifically enabled. However, this could potentially be overlooked when adding future commands and options. The commit introduces a simple test to check that usage messages for all commands and subcommands display default values for any boolean option they display.