Check head state and assignments before attempting sync committee duty.
If a sync committee duty cannot get the current chain head, and there are no assignments, generating an error has limited value, so in the case where this condition is met, treat the duty as a no op.

Also, if there are no assignments, there is no need to subscribe to any subnets.

fixes #4370

Signed-off-by: Paul Harris <paul.harris@consensys.net>