Sync committee duties should not perform aggregates if there are no validators active
There were actually 2 duties performed in sync committee, and the previous PR addressed only 1.

Also don't request attestation or block duties if there are no keys, as the result will be empty, and for remote validators that are services, it is just adding to the request count for no value.

fixes #4370

Signed-off-by: Paul Harris <paul.harris@consensys.net>