Add a cache to track liveness of validators
This cache is specifically needed to service rest api requests, so is not required if we are not using the rest api.

The rest endpoint is yet to be approved, so leaving it marked experimental until the endpoint is finalized.

Currently feature flagged with 'Xbeacon-liveness-tracking-enabled', which is a hidden CLI option.

Enabling will keep a 2d array of as much as 3x (the max validator index+1000) to maintain where validators have been seen.

partially addresses #4453

Signed-off-by: Paul Harris <paul.harris@consensys.net>