Increase delay between pinging bootnodes
Ensures that the delay between pings is longer than the session timeout from discovery so previously failed ping attempts are forgotten and we retry from a clean slate. This is a quick fix, ultimately the session management in discovery needs to be adjusted to drop sessions that are non-responsive faster when they are in the initial handshake phase.