Create a labelled gauge to report connected peers.
Peers will be logged by their client type (eg. teku, lighthouse, prysm) without version number so that the reporting is relatively clean.

Once the initial status is determined, an 'identify' request will be made to the peer, and it will be persisted on the peer object for reporting.

Also stored the peer client type on the peer object, but we don't currently do anything with it, so it could be removed arguably.

fixes #2764

Signed-off-by: Paul Harris <paul.harris@consensys.net>