Default to --network=auto for validator client #4471 (#4629)
This commit changes the default value for the CLI network option from 'mainnet' to 'auto', which fetches relevant details from the configured beacon node endpoint automatically. This change is inline with the deprecation plan for the CLI network option outlined in #3530. Part of #4471

Throws InvalidConfigurationException if --network is specified before the validator-client subcommand which would otherwise cause it to not be applied correctly.