Remove legacy ProtoArraySnapshot (#4682)
Quite some time ago, protoarray contents were stored in a single snapshot which could get out of sync with the blocks in Store. We implemented an automatic migration from that system to the new one where they are stored correctly in-sync.  This removes that migration code and the old ProtoArraySnapshot.  Any user still running a version of Teku old enough to have not migrated away from ProtoArraySnapshot storage won't be compatible with Altair and need to resync anyway. So we can safely remove this migration code now and simplify our startup process.