Reject create block and attestation requests when chain head is optimistic
It is not safe for validators to produce blocks on top of or attest to blocks that have only been optimistically synced. So if the referenced block is not fully validated, return a node syncing response rather than creating the block/attestation. Node syncing is used because the execution engine is still syncing and unable to verify the payload.

Since blocks and attestations use the chain head which is normally not optimistic block production and attestation creation can continue while there are some optimistically sync'd blocks, as long as we still have a fully validated block close enough to use as our chain head and build on. If we've optimistically finalized, we will cease block and attestation production.