Slashing protection import accepts a consumer for logging
Also, any errors are returned as Optional strings, to allow the caller to determine how to show errors.

This will allow an api to use the importer, not just restrict it to being used by the stand-alone import command.

Care should still be taken not to update records for active validators - this was previously not possible but could be possible now. That would need to be looked after by the caller, but the importer will ensure any records are merged to file.

Signed-off-by: Paul Harris <paul.harris@consensys.net>