Perform state transition for blocks off the fork choice thread (#4751)
Since state transition does not require a mutable store, it can be performed off the fork choice thread, then once the post-state is complete, adding the block to the store and protoarray can happen on the fork choice thread.

This simplifies execution of post-merge payloads as the executePayload request is sent to the EL from off the fork choice thread and the CL can then wait for a response prior to switching to the fork choice thread. Thus when the EL is online, blocks are imported as valid immediately and optimistic sync is only used when the EL is actually syncing.