Validator id parameters should not be rejected for bytes48 not on g2 curve
If a bytes48 key is provided that can't be validated on the curve, treat it as a missed key, not an invalid request. In a large list of bytes48 a single key will otherwise reject the entire request, where we could instead just process the rest of the request.

Internally we still need to map keys to the curve to be able to search for the key, so for any valid bytes48 we will still end up having to map to the curve to find the key.

0x hex with invalid length will still be rejected as a bad request.

fixes #4787

Signed-off-by: Paul Harris <paul.harris@consensys.net>