Add authentication to validator rest api.
will create a file in `validator/validator-api-bearer` with a random password.

If the password is "xyz", that is what is in that file, and then the Authorization header is expected to be `Bearer xyz`

The first line of the file is read and then uri encoded to remove any invalid data. For example, ` ` gets encoded to `+`.

Users and roles are possible but not implemented in this version, as we're keeping it simple.

The password is read at startup, so changing password would currently require a restart.

Because of the clear text nature of the password in the header, it's going to need to be changed to https transport.

Signed-off-by: Paul Harris <paul.harris@consensys.net>