Enable and disable gossip based on optimistic sync state.
Moves away from the network periodically polling recentChainData to decide when to enable gossip to explicitly notifying it of sync state changes.
This allows gossip to be reduced to just blocks when optimistic sync is enabled and also allows unsubscribing from gossip when the node falls too far behind (such as when a laptop wakes from sleep).