Add beacon node publish metrics
Implement the publish mechanism for beacon node metrics.

A couple of fields are defaulted but most can be provided. Added a labelled counter for bytes sent/received on libp2p, which isn't monitoring all sent bytes, but will any time we're sending gossip which is probably a very good start.

partially addresses #4391

Signed-off-by: Paul Harris <paul.harris@consensys.net>