Track the execution payload of optimisically sync'd, finalized merge transition block.
So that we can validate the terminal block conditions when the EL is back online, keep a record of the execution payload for the merge transition block if it was optimistically verified and has been finalized. On disk we only store the slot of the merge block and can reload the ExecutionPayload from that.

Not done yet is adding the ability to clear the recorded payload when it has later been verified (or to actually verify it later...)