Start to convert `Map<?, Integer>` to `Object2IntMap<?>`
Adrian mentioned [0] that we could convert these maps to the fastutil version.
I'm a little unsure if `SynchronizedMap` is a good replacement for
`ConcurrentHashMap` but I think it is. Also, not sure if I need to change the
type of `validatorIndexesByPublicKey` in `BeaconProposerPreparerTest.java` too.

There are about 20 more instances of these maps. I plan to convert those too.

[0] https://github.com/ConsenSys/teku/pull/4991#discussion_r808554792