GetBlockHeader should set canonical=false if appropriate
If selecting by block root, you can select a non canonical block, but the header wasn't correctly filling in the canonical boolean.

Added tests for GetHeader, also added a basic integration test for `GetBlockHeaders`, which really needs to be expanded upon so that it can return non canonical headers.

fixes #5080

Signed-off-by: Paul Harris <paul.harris@consensys.net>