Remove EventChannel indirection for time tick events (#5171)
Moves control of the TimerService into BeaconChainController so that it can call the onTick method directly rather than going via an event channel. This removes the queue of events and allows the quartz scheduler to manage back pressure if events aren't processed fast enough. It does this by skipping forward, making it less likely (though still possible) for the beacon chain's view of time to fall behind.