Fix FuzzUtilTests
So all of the tests pass now! This involved replacing all of the ssz files with
their respective bellatrix versions. I needed to run `gen_operations` and
`gen_sanity` in ethereum/consensus-specs to make them.

I needed to modify the `dump_ssz_fn` in gen_runner.py to not snappy compress
the serialized data before writing it to disk. Not sure if this is new or what,
but the test suite does not work with snappy data. To be fair, that's
unnecessary work for fuzzing.

Next, I need to update the yaml files next. I think I'm going to need to modify
the generation functions to tell it to output yaml. Not sure how the original
developers of this did that.