Start moving parameter utils to infrastructure.
We don't want individual handlers to have to handle all of the complexity of dealing with query parameters, so it makes sense to use parameterUtils, but currently they're not accessible.

Also removed the `getQueryParamMap` function, as ideally we won't need to have direct access to the query or path paramter maps from the context.

Signed-off-by: Paul Harris <paul.harris@consensys.net>