Terminal blocks must not be in the future
When selecting the terminal block, ensure that it's timestamp is in the past. Otherwise a block built on it may be invalid because it's timestamp (the slot start time) is before the parent block timestamp.

This is extremely unlikely to occur in normal PoW networks but regularly happens when using low fixed difficulty mining in local test networks.