Initial schema definitions for Event types that aren't just standard objects
These were missing from the schema definitions we use from swagger, this test creates a definition from the class, which is a little ugly.

Once we're using the new method this test will be a lot cleaner, but this should allow us to be confident the structure is correct, despite the lacking typing information.

Signed-off-by: Paul Harris <paul.harris@consensys.net>