Continue polling engine_exchangeConfiguration after the merge completes (#5327)
From the execution engine spec: EL clients SHOULD log an error if they don't receive a call to engine_exchangeConfiguration at least once every 120 seconds. This doesn't include a provision for stopping polling after the merge completes, and the EL may not have the same view of the merge status as we do due to different syncing stages.  So we need to keep polling even after we see the merge as complete.