Add time based tolerance before disconnecting event streams.
Given that huge numbers of attestations are received at once, a strict queue length based limit on event stream subscribers is too strict as there's no chance to serialize and read all those events. With this change, the event queue can exceed the maximum size for up to one second, ensuring there is time for the messages to be sent if the subscriber is reading. If the subscriber is slow the timeout is still hit quickly without too many events backing up and causing memory issues.