Optimise storage of attestations waiting to be applied to fork choice.
Attestations can only be applied to fork choice in the slot after they were created. Previously the whole attestation was stored in the generic future attestation pool. Now the minimal info required to update protoarray is stored in a dedicated structure for attestations that have already been added to the aggregating pool but not applied to fork choice.