Add propertyTest task
This new task will run all of our Property-Based Testing (PBT) tests. With PBT,
you test a certain property by defining constraints and letting the test engine
produce randomized inputs; it's a lot like fuzzing. I discovered this concept
while looking for an equivalent to Golang's `go fuzz` feature [0], introduced
in Go 1.18, that would work with Teku and its dependencies.

These first few property tests are relatively simple. I had a difficult time
decided where to start. I know that this area is already pretty well tested,
most of Teku is pretty well tested IMO, but I needed to start somewhere. The
plan is to add more PBT tests to existing classes in the near future.

The PBT test engine that I chose was not my first choice. Initially, I
evaluated JQF [2] which features randomized testing with coverage-guided
fuzzing algorithms and the ability to run a certain test indefinitly.
Unfortunately, JQF does not work well with Teku. It requires the JUnit4
"vintage" platform and when we use the newer JUnit5 "jupiter" platform. Also,
JQF is pretty tightly integrated with Maven and does not really support Gradle.
There is a Gradle plugin [2] but it's still very much in development; it
doesn't support project structures like Teku.

Then, I evaluated `java-quickcheck`, which JQF is based on. As expected, it had
the same problems as JQF. It doesn't feature coverage-guidance or anything cool
like that but it would have still been useful.

Lastly, I tested `jqwik` which is an alternative PBT test engine that is
designed to work for the JUnit5 platform. In my opinion, it has really great
documentation [3][4] and, from the best I can tell, it has all of the same
features as `java-quickcheck`. Another good quality, there's an additional
`jqwik-kotlin` module that is supposed to simplify/streamline using `jqwik` in
Kotlin projects. This will be useful if we decide to introduce PBT into libp2p
in the future as it is primarily written in Kotlin.

[0] https://go.dev/doc/fuzz/
[1] https://github.com/rohanpadhye/JQF
[2] https://github.com/mborzenkov/jqf-gradle-plugin
[3] https://jqwik.net/docs/current/user-guide.html
[4] https://jqwik.net/docs/1.6.5/javadoc/index.html