Add some JsonProvider property tests
Currently, the roundTripBeaconState test fails. I see this error:

  Invalid Bitvector ssz: trailing bits are not 0

I think this is because it is deserializing the string bytes instead of the hex
representation. In roundTripBitVector, I call Bytes.fromHexString on the string
prior to deserialization. I am unsure how to do this for a serialized beacon
state string.

Also, I needed to add syncAggregate & executionPayload fields to the beacon
block generator. For altair, these are required. The other generator had this
already. I would see this error before:

  java.lang.NullPointerException:
    syncAggregate must be specified