Don't print the exception trace for remote metrics publishing
If the remote metrics service isn't available, stack trace is not helpful.

If the error is a JsonProcessingException, it will still print stack trace as it may be a bug we need context of, or a change in the specification that we need to update our client to handle.

Also, the log message should be warning level for this specific message.

fixes #5416

Signed-off-by: Paul Harris <paul.harris@consensys.net>