Don't mark blocks before latestValidHash as fully validated (#5951)
While latestValidHash guarantees that every block after it is invalid, there are cases where the EL may not have fully executed the block returned as latestValidHash (typically because it's on a non-canonical fork). As such it is not safe to mark those nodes as fully VALID and they should be left as optimsitically imported.