Cleanup moving hot blocks to finalized after restarts for blinded blocks.
 - also added some better exception handling and extra context to migration script

 - added some context to some of the database test data.

To see this in reality, you would need to be using blinded block storage. On restart, it's possible blocks in your 'hot' will go missing, rather than get copied to finalized storage as would be normally expected if they're canonical. You'd also have to be looking for parents of finalized blocks to actually find the locations where there are issues, so it's non trivial to track down, and I'm glad we could test for it.

Signed-off-by: Paul Harris <paul.harris@consensys.net>