Skip finding the PoW block that first satisfies the minimum genesis time condition when we already have an anchor state.
The min genesis time block is only required as part of the process to generate genesis so if we already have a BeaconState to start from, we don't need to generate the genesis state and thus don't need to find this block.

This improves compatibility with ELs (particularly Nethermind) that may not download all historic blocks or may backfill them after eth_syncing reports the node is in sync.