Use the fork from the supplied state when validating exit and slashing signatures
This will mean gossip is validated against the current chain head fork rather than the fork scheduled at the time of the slot, matching other client's behaviour.
The signature check when building a block will then use the fork from the block pre-state which ensures the operation is actually valid.

Also updates the reference test executor for operations to check that we include valid operations and exclude invalid ones when creating a block correctly to improve test coverage.