Support async storage updates.
Adds a feature toggle (default off) to allow updating the in-memory Store without first storing the update to disk.

Currently, newly finalized data may be temporarily unavailable as it may be removed from the in-memory store before it is persisted to disk. Bufferring will be added in a follow up PR.