Preserve ordering of updates and queries when async IO is enabled.
Moves storage updates and queries to use the same event channel when async IO is enabled.
This ensures that any query sent to the channel after an update only begins executing after
the update completes. To avoid updates being delayed by queries, the queries are actually
executed on a separate async runner. This means that queries may be delayed by updates but
not the other way around. It avoids the need to try and buffer content to deal with updates
being asynchronous and finalized data being unavailable until the update completes.