Update reference tests to 1.3.0-alpha.1 (#6484)
Also adjusts sync aggregate processing to ensure safe interactions with withdrawals in the (exceptionally unlikely) case where a fully withdrawn validator appears in the sync committee in two different positions and is penalised for the first position and rewarded for the second. The balance changes must be applied in the exact order as the spec to ensure the min zero condition when reducing the balance has the same results.