Adds sync committee metrics
`validator_scheduled_sync_committee_duties_current`
This new metric is a visualisation to see how many validators are currently assigned to performing sync committee duties.

There were already metrics showing the attestation duties and block production duties, named in the same way, so this was just filling in the gap, although the other duties are accounted in a slightly different way, as they work differently.

This new metric is effectively the number of validators that you have that are in a sync committee currently.

Also added
`validator_current_sync_committee_last_epoch`, which will be the last epoch of the current sync committee.

This will allow validator maintainers to better determine if they are in a sync committee currently, and when the sync committee will be changing.

Fixes #6270

Signed-off-by: Paul Harris <paul.harris@consensys.net>