implement an alternative pruning strategy for smaller block sets to prune. This will require more cpu to read block, but less of a scan, so for small numbers of blocks should be quite clean.
Also adjusted the pruning frequency to keep the numbers small, which is what will make this method practical.

In 15 minutes, should be less than 75 blocks to prune, so the 500 limit for pruning by this new method should easily handle the small number of blocks that trickle in.

Larger numbers can still benefit from using the index, but for smaller scans it's a large overhead, because of the need to scan the entire index.

Signed-off-by: Paul Harris <paul.harris@consensys.net>