Implemented Engine API capabilities provider
This implementation retains all the current functionality, introducing the concept of the Engine API capabilities provider, an interface with a method that returns all available Engine API methods.

The initial implementation LocallySupportedEngineApiCapabilitiesProvider.java basically returns all methods available in Teku.

Also renamed the ExecutionJsonRpcMethodsProvider interface to ExecutionJsonRpcMethodsResolver to better capture the idea of resolving a method from the available capabilities.