Refactor BLS Operation pool for re-use.
Generify the class so that it can be employed by voluntary exit operation pool.

The advantage of this class is that it allows differentiation between locally and remotely submitted operations, which allows us in turn to select preferentially, and potentially in future perform other operations on local objects.

This does substantially increase the potential pool size for voluntary exits, but we can reduce that if we need to.

Partially addresses #6803

Signed-off-by: Paul Harris <paul.harris@consensys.net>