Added sync-committee dependent block fetching
Previously, we may have been ignorant to a new block that may have been slow to arrive via gossip, but now we'll fetch the block referred to by sync committee messages and contributions and proofs.

This is much the same as our pending attestations, where we actively search for blocks we've not seen, then process the gossip once we have the referenced block.

This should ensure we have the best possible quality of sync committee message and contributions and proofs.

Also resolved the spelling of 'Validatable' in a couple of classes.

Signed-off-by: Paul Harris <paul.harris@consensys.net>