Use progressive balances during epoch transition
Also cleaned up some of the processing logic, but only very small tweak.

Basically it's cheap to check if a delta is zero, and if it's zero there's no need to adjust balance. This will occur for every exited validator that has 0 balance, and any exited validator outside of any slashing period, so it's actually about 15% of the validator set, at which point it becomes quite worthwhile to check if the delta is zero, as we save a read+write to the balance set.

This is not a huge speed improvement, but it does round out the progressive balances change.

Fixes #6998

Signed-off-by: Paul Harris <paul.harris@consensys.net>