Added epoch state cache configuration
Now that attestations aren't being validated against states so often, we need a cache closer aligned to what we're doing when we need to regenerate - basically we want to avoid regeneration of state if at all possible.

One epoch transition the finalized and sometimes justified state get read, and were causing regenerations, so adding a cache to handle storing those, but leaving it size 0 by default for now.

The idea would be to have the current state cache reduced significantly in size, potentially to about half an epoch or so, and have an epoch cache with a maximum size over 2, and could probably be below 10 by default.

This would mean the vast majority of the time, no state regeneration is required, and our states in memory can drop to a much lower number than currently stored in memory (32 * 5 max by default currently).

Tested in minimal network, needs to be deployed to a more complex setup to see if it behaves well in a real environment.

Signed-off-by: Paul Harris <paul.harris@consensys.net>