increased default to suite most database tests
Most database tests had plenty of space with a state cache of 8, and the default of 6 was arbitrary, so I increased the default to 8, and updated notes.

A small set of tests needed 2 epochs of state cache, and it's easy to adjust, so I took that as the path of least resistance, but we might revisit that at some point.

Signed-off-by: Paul Harris <paul.harris@consensys.net>