Updated the default storage mode for new databases to minimal
Previously initialised databases will remain on prune mode if their storage is not specified.

Makes use of the new data-storage-mode file to allow teku to remember the last storage requested on the subsequent opening.

Can still be specified by data-storage-mode and the storage mode will be updated if this is used.

Fixes #7392

Signed-off-by: Paul Harris <paul.harris@consensys.net>