Added a feature to remove keys on reload
Behind a feature toggle. Removes keys in the same way that the keymanager does, to ensure that the signer can't be referenced once the key has been removed.

Made a script that allows removal of keys, so that i can add, reload, remove, reload with that interop folder.

Offers a solution to #7503

Signed-off-by: Paul Harris <paul.harris@consensys.net>