Updated holesky forks and genesis time
We were accepting the current genesis due to some fork checks being skipped during startup, because we detected pre-genesis. I've altered fork checking to always occur, which did mean a few more changes than just holesky configuration.

The upside is we can now configure default initial state to finalized, then genesis, and it'll warn and try the genesis state if the finalized fails, and it'll determine if the forks change (like on holesky) and error.
example output:
```
13:56:52.665 INFO  - Loading initial state from https://checkpoint-sync.holesky.ethpandaops.io/eth/v2/debug/beacon/states/finalized
13:56:53.006 WARN  - Failed to load initial state from https://checkpoint-sync.holesky.ethpandaops.io/eth/v2/debug/beacon/states/finalized
13:56:53.006 INFO  - Loading initial state from https://checkpoint-sync.holesky.ethpandaops.io/eth/v2/debug/beacon/states/genesis
13:57:11.233 INFO  - Loaded initial state at epoch 0 (state root = 0x706eeb43c9e44932b312cd26b8d35ef69725f117275ef3dcd67653625b519f16, block root = 0x00af3346d685d33731a476469262112fee02e82b05e60d83c577c649b33ad152, block slot = 0).
The fork from the intial-state (0:0x20017000) does not match the configured fork schedule (0:0x03017000).
Please check that network in configuration matches the loaded state.
```

We could still reference the genesis state from github, but it's not currently merged.

Partially addresses #7517

Signed-off-by: Paul Harris <paul.harris@consensys.net>